+++ packages/gestalt/src/Masonry/__integration__/addRelatedItems.ghost.js
@@ -23,7 +23,7 @@ const getItemColumnMap = async () => {
 describe('Masonry > Add related items', () => {
   it('Can insert related items into the grid', async () => {
     ghost.close();
-    await ghost.open('http://localhost:3000/Masonry', {
+    await ghost.open('http://localhost:3001/Masonry', {
       viewportSize: {
         width: 3000,
         height: 2000,