+++ CHANGELOG.md
@@ -6,6 +6,24 @@
     Click to see more.
   </summary>
 
+  ### Minor
+
+  ### Patch
+  * Internal: Fix publish script to work for new Gestalt directory structure (#94)
+
+</details>
+
+## 0.62.0 (March 21, 2018)
+
+### Minor
+* Heading / Text / SegmentedControl: Add `title` when `truncate` is set (#82)
+
+### Patch
+* Docs: Masonry locally on port `3000` + update the `README` with the latest commands (#89)
+* Internal: No downtime when releasing the docs (#97)
+
+## 0.61.0 (March 20, 2018)
+
 ### Minor
 
 * Image: Don't show `alt` text when loading the image in FireFox. (#80)(#84)
@@ -16,8 +34,6 @@
 * Docs: Add live docs to Toast (#87)
 * Internal: Convert `BrowserRouter` to `HashRouter` - fixes directly going to a component (#88)
 * Docs: Add live docs to SegmentedControl (#90)
-
-</details>
 
 ## 0.60.0 (March 13, 2018)